+++ drone_rew/hover_train.py
@@ -1,3 +1,5 @@
+import faulthandler
+faulthandler.enable()
 import argparse
 import os
 import pickle
@@ -29,10 +31,10 @@ def get_train_cfg(exp_name, max_iterations):
             "entropy_coef": 0.004,
             "gamma": 0.99,
             "lam": 0.95,
-            "learning_rate": 0.0003,
+            "learning_rate": 0.0001,
             "max_grad_norm": 1.0,
-            "num_learning_epochs": 5,
-            "num_mini_batches": 4,
+            "num_learning_epochs": 6,
+            "num_mini_batches": 64,
             "schedule": "adaptive",
             "use_clipped_value_loss": True,
             "value_loss_coef": 1.0,
@@ -40,8 +42,8 @@ def get_train_cfg(exp_name, max_iterations):
         "init_member_classes": {},
         "policy": {
             "activation": "tanh",
-            "actor_hidden_dims": [128, 128],
-            "critic_hidden_dims": [128, 128],
+            "actor_hidden_dims": [256, 128],
+            "critic_hidden_dims": [256, 128],
             "init_noise_std": 1.0,
             "class_name": "ActorCritic",
         },
@@ -57,9 +59,9 @@ def get_train_cfg(exp_name, max_iterations):
             "run_name": "",
         },
         "runner_class_name": "OnPolicyRunner",
-        "num_steps_per_env": 100,
-        "save_interval": 100,
-        "empirical_normalization": None,
+        "num_steps_per_env": 400,
+        "save_interval": 50,
+        "empirical_normalization": True,
         "seed": 1,
     }
 
@@ -70,17 +72,17 @@ def get_cfgs():
     env_cfg = {
         "num_actions": 4,
         # termination
-        "termination_if_roll_greater_than": 180,  # degree
-        "termination_if_pitch_greater_than": 180,
-        "termination_if_close_to_ground": 0.1,
-        "termination_if_x_greater_than": 3.0,
-        "termination_if_y_greater_than": 3.0,
+        "termination_if_roll_greater_than": 90,  # degree
+        "termination_if_pitch_greater_than": 90,
+        "termination_if_close_to_ground": 0.3,
+        "termination_if_x_greater_than": 5.0,
+        "termination_if_y_greater_than": 5.0,
         "termination_if_z_greater_than": 2.0,
         # base pose
-        "base_init_pos": [0.0, 0.0, 1.0],
+        "base_init_pos": [0.0, 0.0, 2.0],
         "base_init_quat": [1.0, 0.0, 0.0, 0.0],
-        "episode_length_s": 15.0,
-        "at_target_threshold": 0.1,
+        "episode_length_s": 60.0,
+        "at_target_threshold": 0.05,
         "resampling_time_s": 3.0,
         "simulate_action_latency": True,
         "clip_actions": 1.0,
@@ -88,30 +90,45 @@ def get_cfgs():
         "visualize_target": False,
         "visualize_camera": False,
         "max_visualize_FPS": 60,
+        #success timer
+        "success_hold_timesteps": 30, #number of timesteps to hold the drone at the target position for a successful episode
+        #domain randomization
+        "domain randomization": {
+             "enabled": True,
+             "mass_scale": [0.95, 1.05],      # Randomly scale mass
+             "actuator_scale": [1.0, 1.0],
+             "action_smoothing_scale": [1.0, 1.0], # Randomly scale action smoothing (inertia) 
+        }
     }
     obs_cfg = {
         "num_obs": 17,
+        "add_noise": True,
+        "noise_level": 0.01,
         "obs_scales": {
             "rel_pos": 1 / 3.0,
-            "lin_vel": 1 / 3.0,
+            "lin_vel": 1 / 8.0,
             "ang_vel": 1 / 3.14159,
         },
     }
     reward_cfg = {
         "yaw_lambda": -10.0,
+        "target_lambda": -6,
         "reward_scales": {
             "target": 10.0,
             "smooth": -1e-4,
-            "yaw": 0.01,
-            "angular": -2e-4,
+         #   "yaw": 0.01,
+         #   "angular": -2e-4,
             "crash": -10.0,
+            "go_near_target": 2.0,
+            "target_completion": 20.0,
+         #  "stay_on_target": 10.0,
         },
     }
     command_cfg = {
         "num_commands": 3,
-        "pos_x_range": [-1.0, 1.0],
-        "pos_y_range": [-1.0, 1.0],
-        "pos_z_range": [1.0, 1.0],
+        "pos_x_range": [-4, 4],
+        "pos_y_range": [-4, 4],
+        "pos_z_range": [1.0, 3.0],
     }
 
     return env_cfg, obs_cfg, reward_cfg, command_cfg
@@ -123,6 +140,8 @@ def main():
     parser.add_argument("-v", "--vis", action="store_true", default=False)
     parser.add_argument("-B", "--num_envs", type=int, default=8192)
     parser.add_argument("--max_iterations", type=int, default=301)
+    parser.add_argument("--resume", action="store_true", default=False, help="Resume training from a checkpoint")
+    parser.add_argument("--load_model", type=str, default=None, help="Path to the model checkpoint to load")
     args = parser.parse_args()
 
     gs.init(logging_level="warning")
@@ -131,17 +150,22 @@ def main():
     env_cfg, obs_cfg, reward_cfg, command_cfg = get_cfgs()
     train_cfg = get_train_cfg(args.exp_name, args.max_iterations)
 
-    if os.path.exists(log_dir):
-        shutil.rmtree(log_dir)
-    os.makedirs(log_dir, exist_ok=True)
+    if not args.resume and os.path.exists(log_dir):
+        if os.path.exists(log_dir):
+            shutil.rmtree(log_dir)
+        os.makedirs(log_dir, exist_ok=True)
 
     if args.vis:
         env_cfg["visualize_target"] = True
 
-    pickle.dump(
-        [env_cfg, obs_cfg, reward_cfg, command_cfg, train_cfg],
-        open(f"{log_dir}/cfgs.pkl", "wb"),
-    )
+    if not args.resume:
+        pickle.dump(
+            [env_cfg, obs_cfg, reward_cfg, command_cfg, train_cfg],
+            open(f"{log_dir}/cfgs.pkl", "wb"),
+        )
+    if args.resume:
+        #env_cfg, obs_cfg, reward_cfg, command_cfg, _ = pickle.load(open(f"{log_dir}/cfgs.pkl", "rb"))
+        train_cfg = get_train_cfg(args.exp_name, args.max_iterations)
 
     env = HoverEnv(
         num_envs=args.num_envs,
@@ -153,7 +177,10 @@ def main():
     )
 
     runner = OnPolicyRunner(env, train_cfg, log_dir, device=gs.device)
-
+    if args.resume:
+            resume_path = args.load_model
+            print(f"Resuming training from checkpoint: {resume_path}")
+            runner.load(resume_path)
     runner.learn(num_learning_iterations=args.max_iterations, init_at_random_ep_len=True)
